Refactor email template generation to be a management command
