+++ docker-entrypoint.sh
@@ -22,8 +22,7 @@ if [[ "$CREATE_ADMIN_USER" = "1" ]]; then
     fi
 fi
 
-python ./scripts/load_notification_templates.py
-echo "Notification templates loaded"
+python /app/manage.py generate_notifications_from_templates
 
 # Start server
 if [[ ! -z "$@" ]]; then
